+++ changelog.md
@@ -30,6 +30,10 @@
 
 # Unreleased
 
+# 1.0.4
+
+- Fix save with Sqlite Storage
+
 # 1.0.3
 
 - Fix drawer load